+++ js/util.js
@@ -1,0 +1,23 @@
+import './util.js';
+
+const getRandomNumberSimple = (numberFrom, numberNext) => {
+  if (numberFrom >= 0 && numberNext >= 0 && numberFrom < numberNext) {
+    const numberMin = Math.ceil(numberFrom);
+    const numberMax = Math.floor(numberNext);
+    const randomSimple = Math.floor(Math.random() * (numberMax - numberMin + 1)) + numberMin;
+    return Number(randomSimple);
+  }
+  throw new TypeError('Некоректно указанны данные');
+};
+
+const getRandomNumberFloat = (numberFrom, numberNext, numberFloat = 0) => {
+  if (numberFrom >= 0 && numberNext >= 0 && numberFrom < numberNext && numberFloat >= 0) {
+    const randomFloat = (Math.random() * (numberNext - numberFrom) + numberFrom).toFixed(numberFloat);
+    return Number(randomFloat);
+  }
+  throw new TypeError('Некоректно указанны данные');
+};
+
+const getRandomElement = (element) => element[getRandomNumberSimple(0, element.length - 1)];
+
+export {getRandomNumberSimple, getRandomNumberFloat, getRandomElement};
